Implement media_path in builders

+++ src/disseminate/builders/scalesvg.py
@@ -35,7 +35,7 @@ class ScaleSvg(Builder):
     scale = None
 
     def __init__(self, env, scale, **kwargs):
-        self.scale = int(scale)
+        self.scale = float(scale)
 
         super().__init__(env, **kwargs)
